Remove the api:: module and move modules around.
