--- luminance/src/api.rs
@@ -1,11 +1,0 @@
-//! API.
-
-pub mod buffer;
-pub mod framebuffer;
-pub mod pipeline;
-pub mod render_gate;
-pub mod shader;
-pub mod shading_gate;
-pub mod tess;
-pub mod tess_gate;
-pub mod texture;
